Moved 5 tests from PileTests to StockPileTests because they all deal specifically with the stock pile (not piles in general).
